pyproject.toml
[project]
name = "fastapitemplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "greenlet>=3.1.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "pytz>=2025.2",
    "redis>=6.1.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]
